move types to dedicated file
